Change int to long

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -1,15 +1,15 @@
 #include <stdio.h>
 
 /**
- * main - Entry point
+ * main - Entry polong
  *
  * Return: Always 0.
  */
-int main(void)
+long main(void)
 {
-	int i;
-	int n1, n2;
-	int sum;
+	long i;
+	long n1, n2;
+	long sum;
 
 	n1 = 0;
 	n2 = 1;
@@ -18,13 +18,13 @@ int main(void)
 	{
 		sum = n1 + n2;
 
-		printf("%d, ", sum);
+		printf("%ld, ", sum);
 
 		n1 = n2;
 		n2 = sum;
 	}
 
-	printf("%d\n", (n1 + n2));
+	printf("%ld\n", (n1 + n2));
 
 	return (0);
 }
